requirements.txt
numpy==1.26.4
asgiref==3.9.1
Django==5.2.6
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
celery==5.4.0
redis==5.0.7
pydicom==2.4.4
Pillow==10.4.0
requests==2.32.3
djangorestframework==3.14.0
django-cors-headers==4.0.0
tensorflow==2.16.2
opencv-python-headless==4.10.0.84
scikit-image==0.24.0
matplotlib==3.9.1
scipy==1.14.0